fix: orient the capture area after making geometry valid
